Move implementations to the relevant modules
