+++ common/nodejs-utils/src/bin/generate_inventory.rs
@@ -1,9 +1,9 @@
 #![warn(clippy::pedantic)]
 
-use heroku_nodejs_utils::{distribution::Distribution, inv::BUCKET};
+use heroku_nodejs_utils::{distribution::Distribution, distribution::DEFAULT_BUCKET};
 
 fn main() {
-    let bucket = std::env::var("AWS_S3_BUCKET").unwrap_or(BUCKET.to_string());
+    let bucket = std::env::var("AWS_S3_BUCKET").unwrap_or(DEFAULT_BUCKET.to_string());
     let inventory = std::env::args()
         .nth(1)
         .unwrap_or_else(|| {
