+++ common/nodejs-utils/src/distribution.rs
@@ -1,12 +1,16 @@
 use crate::{
-    inv::{Inventory, Release, BUCKET, REGION},
+    inv::Inventory,
     nodejs_org, npmjs_org, s3,
-    vrs::{Requirement, Version},
+    vrs::{Requirement, Version, VersionSet},
 };
 use anyhow::anyhow;
 use regex::Regex;
-use std::convert::TryFrom;
-use std::{collections::HashSet, fmt, str::FromStr};
+use std::{fmt, str::FromStr};
+
+/// Heroku nodebin AWS S3 Bucket name
+pub const DEFAULT_BUCKET: &str = "heroku-nodebin";
+/// Heroku nodebin AWS S3 Region
+pub const DEFAULT_REGION: &str = "us-east-1";
 
 #[derive(Debug, Clone, Copy)]
 pub enum Distribution {
@@ -53,7 +57,7 @@ impl Distribution {
     /// - http call failures
     /// - http response parsing errors
     pub fn mirrored_versions(&self, bucket: &str) -> anyhow::Result<VersionSet> {
-        s3::list_objects(bucket, REGION, self.bucket_prefix())?.try_into()
+        s3::list_objects(bucket, DEFAULT_REGION, self.bucket_prefix())?.try_into()
     }
 
     /// Build an inventory from the releases listed in the mirrored location.
@@ -62,7 +66,7 @@ impl Distribution {
     /// - http call failures
     /// - http response parsing errors
     pub fn mirrored_inventory(&self, bucket: &str) -> anyhow::Result<Inventory> {
-        s3::list_objects(bucket, REGION, self.bucket_prefix())?.try_into()
+        s3::list_objects(bucket, DEFAULT_REGION, self.bucket_prefix())?.try_into()
     }
 
     /// Filter inactive distribution versions and return as a `VersionSet`.
@@ -90,15 +94,14 @@ impl Distribution {
         })
         .map_err(|e| anyhow!("{e}"))
     }
-    fn mirrored_path_regex(self) -> anyhow::Result<Regex> {
+
+    pub(crate) fn mirrored_path_regex(self) -> anyhow::Result<Regex> {
         Regex::new(match self {
             Self::Node => r"node/(?P<channel>\w+)/(?P<arch>[\w-]+)/node-v(?P<version>\d+\.\d+\.\d+)[\w-]+\.tar\.gz",
             Self::Yarn => r"yarn/(?P<channel>\w+)/yarn-v(?P<version>\d+\.\d+\.\d+(-[\w\.]+)?)\.tar\.gz",
         }).map_err(|e| anyhow!("Mirrored release regex error: {e}"))
     }
 }
-
-pub type VersionSet = HashSet<Version>;
 
 fn list_upstream_node_versions() -> anyhow::Result<VersionSet> {
     nodejs_org::list_releases()?
@@ -121,175 +124,9 @@ fn list_upstream_yarn_versions() -> anyhow::Result<VersionSet> {
     Ok(vset)
 }
 
-impl TryFrom<s3::BucketContent> for VersionSet {
-    type Error = anyhow::Error;
-
-    /// # Failures
-    /// These are the possible errors that can occur when calling this function:
-    ///
-    /// * Regex missing matching captures against `Content#key`
-    /// * `Version::parse` fails to parse the version found in the `Content#key`
-    fn try_from(result: s3::BucketContent) -> Result<Self, Self::Error> {
-        let rex = result
-            .prefix
-            .parse::<Distribution>()?
-            .mirrored_path_regex()?;
-
-        result
-            .contents
-            .iter()
-            .map(|content| {
-                rex.captures(&content.key)
-                    .ok_or(anyhow!(
-                        "Couldn't match the bucket content key to a known format: {}",
-                        content.key
-                    ))
-                    .and_then(|capts| {
-                        capts.name("version").ok_or(anyhow!(
-                            "Couldn't find a version number in the bucket content key: {}",
-                            content.key
-                        ))
-                    })
-                    .and_then(|vrs_match| {
-                        Version::parse(vrs_match.as_str()).map_err(|e| {
-                            anyhow!("Couldn't serialize bucket content key as a Version: {e}")
-                        })
-                    })
-            })
-            .collect()
-    }
-}
-
-impl TryFrom<s3::BucketContent> for Inventory {
-    type Error = anyhow::Error;
-
-    /// # Failures
-    /// These are the possible errors that can occur when calling this function:
-    ///
-    /// * Regex missing matching captures against `Content#key`
-    /// * `Version::parse` fails to parse the version found in the `Content#key`
-    fn try_from(result: s3::BucketContent) -> Result<Self, Self::Error> {
-        let dist: Distribution = result.prefix.parse()?;
-        let rex = dist.mirrored_path_regex()?;
-
-        let releases: anyhow::Result<Vec<Release>> = result
-            .contents
-            .iter()
-            .map(|content| {
-                let capture = rex.captures(&content.key).ok_or_else(|| {
-                    anyhow!("No valid version found in content: {}", &content.key)
-                })?;
-                let channel = capture.name("channel").ok_or_else(|| {
-                    anyhow!("Could not find channel in content: {}", &content.key)
-                })?;
-                let version_number = capture.name("version").ok_or_else(|| {
-                    anyhow!("Could not find version in content: {}", &content.key)
-                })?;
-                let arch = capture.name("arch");
-
-                Ok(Release {
-                    arch: arch.map(|a| a.as_str().to_string()),
-                    version: Version::parse(version_number.as_str())?,
-                    channel: channel.as_str().to_string(),
-                    // Amazon S3 returns a quoted string for ETags
-                    etag: Some(content.etag.replace('\"', "")),
-                    url: format!(
-                        "https://{BUCKET}.s3.{REGION}.amazonaws.com/{}",
-                        &content.key
-                    ),
-                })
-            })
-            .collect();
-
-        Ok(Self {
-            name: dist.to_string(),
-            releases: releases?,
-        })
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use chrono::Utc;
-
-    #[test]
-    fn it_converts_s3_result_to_inv() {
-        let etag = "739c200ca266266ff150ad4d89b83205";
-        let content = s3::Content {
-            key: "node/release/darwin-x64/node-v0.10.0-darwin-x64.tar.gz".to_string(),
-            etag: etag.to_string(),
-            ..Default::default()
-        };
-        let bucket_content = s3::BucketContent {
-            prefix: "node".to_string(),
-            contents: vec![content],
-        };
-
-        let result = Inventory::try_from(bucket_content);
-        assert!(result.is_ok());
-        if let Ok(inv) = result {
-            assert_eq!(Some(String::from(etag)), inv.releases[0].etag);
-        }
-    }
-
-    #[test]
-    fn it_converts_s3_result_to_inv_arch_optional() {
-        let content = s3::Content {
-            key: "yarn/release/yarn-v0.16.0.tar.gz".to_string(),
-            ..Default::default()
-        };
-        let bucket_content = s3::BucketContent {
-            prefix: "yarn".to_string(),
-            contents: vec![content],
-        };
-
-        let result = Inventory::try_from(bucket_content);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn it_fails_to_convert_s3_result_to_inv() {
-        let content = s3::Content {
-            key: "garbage".to_string(),
-            last_modified: Utc::now(),
-            etag: "\"e4cc76bea92fabb664edadc4db14a8f2\"".to_string(),
-            size: 7_234_362,
-            storage_class: "STANDARD".to_string(),
-        };
-        let bucket_content = s3::BucketContent {
-            prefix: "yarn".to_string(),
-            contents: vec![content],
-        };
-
-        let result = Inventory::try_from(bucket_content);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn try_from_bucket_content_for_version_set_succeeds() {
-        let content = s3::Content {
-            key: "node/release/darwin-x64/node-v18.10.2-darwin-x64.tar.gz".to_string(),
-            last_modified: Utc::now(),
-            etag: "123abcdefg".to_string(),
-            size: 4_065_868,
-            storage_class: "STANDARD".to_string(),
-        };
-        let bucket_content = s3::BucketContent {
-            prefix: "node".to_string(),
-            contents: vec![content],
-        };
-
-        let vrs_set = VersionSet::try_from(bucket_content)
-            .expect("Expected to convert bucket content to a version set");
-        println!("vrs_set: {vrs_set:?}");
-
-        let expected = Version::parse("18.10.2").expect("Expected to parse a valid version");
-        let actual = vrs_set
-            .get(&expected)
-            .expect("Expected to find a matching version");
-        assert_eq!(&expected, actual);
-    }
 
     #[test]
     fn test_list_upstream_node_versions() {
